==> pipeline/05.VC.R <==
remove(list=ls())

load("./fulldatamodels/fm.avg.rda")
load("./fulldatamodels/fm.filt.rda")
load("./fulldatamodels/fm.g.rda")
load("./fulldatamodels/fm.rda")
load("./fulldatamodels/fm.stg.rda")

fm$varE
fm.conv$varE
fm.filt.uc$varE
fm.g$varE
fm.stages$varE

fm$ETA[[1]]$varU
fm.conv$ETA[[1]]$varU
fm.filt.uc$ETA[[1]]$varU
fm.g$ETA[[1]]$varU
fm.stages$ETA[[1]]$varU

fm$ETA[[2]]$varU
fm.conv$ETA[[2]]$varU
fm.filt.uc$ETA[[2]]$varU
fm.g$ETA[[2]]$varU
fm.stages$ETA[[2]]$varU

fm$ETA[[3]]$varU
fm.conv$ETA[[3]]$varU
fm.filt.uc$ETA[[3]]$varU
fm.g$ETA[[3]]$varU
fm.stages$ETA[[3]]$varU

fm$ETA[[4]]$varU
fm.conv$ETA[[4]]$varU
fm.filt.uc$ETA[[4]]$varU
fm.stages$ETA[[4]]$varU
